move code to __init__.py where appropriate
