+++ frc_characterization/__main__.py
@@ -1,7 +1,7 @@
 # Shim to run the CLI entry point as an ordinary python module for those
 # who are having trouble adding python scripts to PATH.
 
-import frc_characterization.cli.cli as cli
+import frc_characterization.cli as cli
 
 def main():
